delete block pool
