+++ blockchain/api/block_api_test.go
@@ -1,25 +1,31 @@
 package api_test
 
 import (
-	"github.com/it-chain/it-chain-Engine/blockchain"
-	"testing"
-	"github.com/it-chain/it-chain-Engine/blockchain/api"
-	"github.com/magiconair/properties/assert"
-	"github.com/it-chain/midgard"
-	"fmt"
 	"errors"
+	"fmt"
+	"testing"
+
+	"github.com/it-chain/it-chain-Engine/blockchain"
+	"github.com/it-chain/it-chain-Engine/blockchain/api"
+	"github.com/it-chain/midgard"
+	"github.com/magiconair/properties/assert"
 )
 
 type MockBlockQueryApi struct {
 	GetLastBlockFunc func() (blockchain.Block, error)
 }
+
 func (br MockBlockQueryApi) GetLastBlock() (blockchain.Block, error) {
 	return br.GetLastBlockFunc()
 }
-func (br MockBlockQueryApi) GetBlockByHeight(blockHeight uint64) (blockchain.Block, error) { return nil, nil }
+func (br MockBlockQueryApi) GetBlockByHeight(blockHeight uint64) (blockchain.Block, error) {
+	return nil, nil
+}
 func (br MockBlockQueryApi) GetBlockBySeal(seal []byte) (blockchain.Block, error) { return nil, nil }
 func (br MockBlockQueryApi) GetBlockByTxID(txid string) (blockchain.Block, error) { return nil, nil }
-func (br MockBlockQueryApi) GetTransactionByTxID(txid string) (blockchain.Transaction, error) { return nil, nil }
+func (br MockBlockQueryApi) GetTransactionByTxID(txid string) (blockchain.Transaction, error) {
+	return nil, nil
+}
 
 type MockEventRepository struct {
 	LoadFunc func(aggregate midgard.Aggregate) error
@@ -28,19 +34,19 @@ type MockEventRepository struct {
 func (er MockEventRepository) Load(aggregate midgard.Aggregate, aggregateID string) error {
 	return er.LoadFunc(aggregate)
 }
-func (er MockEventRepository) Save(aggregateID string, events ...midgard.Event) error {return nil}
-func (er MockEventRepository) Close() {}
+func (er MockEventRepository) Save(aggregateID string, events ...midgard.Event) error { return nil }
+func (er MockEventRepository) Close()                                                 {}
 
 func TestBlockApi_AddBlockToPool(t *testing.T) {
-	tests := map[string] struct {
+	tests := map[string]struct {
 		input struct {
 			block blockchain.Block
 		}
-	} {
+	}{
 		"success": {
 			input: struct {
 				block blockchain.Block
-			} {block: &blockchain.DefaultBlock{
+			}{block: &blockchain.DefaultBlock{
 				Height: uint64(11),
 			}},
 		},
@@ -64,24 +70,23 @@ func TestBlockApi_AddBlockToPool(t *testing.T) {
 	}
 }
 
-
 func TestBlockApi_CheckAndSaveBlockFromPool(t *testing.T) {
-	tests := map[string] struct {
+	tests := map[string]struct {
 		input struct {
 			height blockchain.BlockHeight
 		}
 		err error
-	} {
+	}{
 		"success": {
 			input: struct {
 				height blockchain.BlockHeight
-			}{height: blockchain.BlockHeight(12),},
+			}{height: blockchain.BlockHeight(12)},
 			err: nil,
 		},
 		"block nil test": {
 			input: struct {
 				height blockchain.BlockHeight
-			}{height: blockchain.BlockHeight(144),},
+			}{height: blockchain.BlockHeight(144)},
 			err: api.ErrNilBlock,
 		},
 	}
