+++ blockchain/block_query_api.go
@@ -1,0 +1,9 @@
+package blockchain
+
+// interface of api gateway query api
+type BlockQueryApi interface {
+	GetBlockByHeight(blockHeight uint64) (Block, error)
+	GetBlockBySeal(seal []byte) (Block, error)
+	GetBlockByTxID(txid string) (Block, error)
+	GetLastBlock() (Block, error)
+}
