+++ blockchain/infra/adapter/command_handler_test.go
@@ -2,15 +2,17 @@ package adapter_test
 
 import (
 	"testing"
-	"github.com/it-chain/midgard"
+
 	"github.com/it-chain/it-chain-Engine/blockchain"
 	"github.com/it-chain/it-chain-Engine/blockchain/infra/adapter"
+	"github.com/it-chain/midgard"
 	"github.com/stretchr/testify/assert"
 )
 
 type MockBlockApi struct {
 	AddBlockToPoolFunc func(block blockchain.Block) error
 }
+
 func (api MockBlockApi) AddBlockToPool(block blockchain.Block) error {
 	return api.AddBlockToPoolFunc(block)
 }
@@ -20,16 +22,16 @@ func (api MockBlockApi) CheckAndSaveBlockFromPool(height blockchain.BlockHeight)
 }
 
 func TestCommandHandler_HandleConfirmBlockCommand(t *testing.T) {
-	tests := map[string] struct {
+	tests := map[string]struct {
 		input struct {
 			command blockchain.ConfirmBlockCommand
 		}
 		err error
-	} {
+	}{
 		"success": {
 			input: struct {
 				command blockchain.ConfirmBlockCommand
-			} {
+			}{
 				command: blockchain.ConfirmBlockCommand{
 					CommandModel: midgard.CommandModel{ID: "zf"},
 					Block: &blockchain.DefaultBlock{
@@ -42,10 +44,10 @@ func TestCommandHandler_HandleConfirmBlockCommand(t *testing.T) {
 		"block nil error test": {
 			input: struct {
 				command blockchain.ConfirmBlockCommand
-			} {
+			}{
 				command: blockchain.ConfirmBlockCommand{
 					CommandModel: midgard.CommandModel{ID: "zf"},
-					Block: nil,
+					Block:        nil,
 				},
 			},
 			err: adapter.ErrBlockNil,
