+++ blockchain/infra/adapter/command_service.go
@@ -2,6 +2,7 @@ package adapter
 
 import (
 	"errors"
+
 	"github.com/it-chain/it-chain-Engine/blockchain"
 	"github.com/it-chain/midgard"
 	"github.com/rs/xid"
